requirements.txt
django==2.0.3
jsonfield==2.0.2
djangorestframework==3.7.7
django_extensions==2.0.6
google-cloud-vision==0.30.1
